Implement the deserialization of tagged values
Add a new method called `visit_tagged_value` to the Visitor trait.
The method has a default implementation deserializing just the value
for all formats without tags.

NOTE: This is not true yet, the default implementation is missing atm.

A new trait `Tagger` is added, that is used to hold information about
the tag.